+++ src/clj_kondo/impl/cache.clj
@@ -67,9 +67,7 @@
       (let [file (cache-file cache-dir lang ns-sym)
             _ (io/make-parents file)
             os (io/output-stream file)]
-        (with-open [;; first we write to a baos as a workaround for transit-clj #43
-                    ;; bos (java.io.ByteArrayOutputStream. 1024)
-                    os (no-flush-output-stream os)]
+        (with-open [os (no-flush-output-stream os)]
           (let [writer (transit/writer os :json)]
             (transit/write writer ns-data)))))))