+++ corpus/my/project/foo.clj
@@ -1,0 +1,3 @@
+(ns my.project.foo)
+
+(defn- bar [])