+++ CHANGELOG.md
@@ -25,6 +25,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - [#2152](https://github.com/clj-kondo/clj-kondo/issues/2152): Fix false positive with used-underscored-binding with core.match
 - [#2150](https://github.com/clj-kondo/clj-kondo/issues/2150): allow command line options = as in --fail-level=error
 - [#2149](https://github.com/clj-kondo/clj-kondo/issues/2149): `:lint-as clojure.core/defmacro` should suppress `&env` as unresolved symbol
+- [#2111](https://github.com/clj-kondo/clj-kondo/issues/2111):  Wrong type on `clojure.core/zero?`
 
 ## 2023.07.13